New utility macros and functions
These should remove the need for most of the manual checks currently
duplicated in the code. Notably the fits_open_file function is defined
as a macro function, so it cannot be passed around and therefore cannot
be used with our fits_invoke function; hence we need to handle it
separately.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>