New fits_file class
The new class takes ownership of a fitsfile pointer and closes it
automatically upon destruction. This ensures all files are closed, even
when exceptions are thrown.

To have this class work efficiently with the existing code, type
conversion operators have been added to fitsfile* and fitsfile**.
Ownership is taken either by receiving a fitsfile* at construction time,
or via assignment.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>