Use safer type for automatic file closing
The Cfits_read_img/Cfits_read_img_subset functions were opening the fits
file using the raw fits_open_file routines, and therefore required
manually closing the file as well. This was done in most cases in
Cfits_read_img, except when an unsupported data type was found, in which
case the file wasn't properly closed. In the case of
Cfits_read_img_subset the file was never closed.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>