Overcoming weird R CMD config behavior
`R CMD config` doesn't output error information to stderr; instead it
outputs it to stdout, which in turn makes direct storage of stdout not
perform as we expected to. We hoped that an error would result in an
empty stdout, and therefore an empty variable; instead one needs to
check the return code of the process and act accordingly.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>