Change return type to void to avoid UB
Not returning a value form a function that specifies a return value type
is undefined behavior, and thus can lead to different kinds of problems
depending on compiler, optimisation levels, etc. The
Cfits_write_img_subset function is not meant to return anything, and
therefore should be marked as returning void.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>