Avoid initialising to-be-written vectors
Initialising and not writing into these new vectors means we not only
avoid writing twice into them, but more importantly we don't produce all
the page faults in serial during initialisation, but potentially in
parallel while reading the FITS file. In systems that don't have "huge
page" support enabled, this makes a huge difference in runtime, with 3.2
GB images taking well above a second in serial, and a fraction of a
second when this happens in parallel.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>