Add multi-threaded subimage reading routine
These are analogous to the multi-threaded full-image routines, except
that instead of adjusting the start/count parameters we need to adjust
the fpixel/lpixel parameters that are sent to the underlying cfitsio
functions. In principle we would want to always parallelise on the
outmost dimension with multiplicity bigger than 1, but for the time
being this implementation only caters for 2D image subset reading.
Adjusting to 3D and 4D image subset reading is left for the future.

This commit introduces additionally introduces the same improvements
already introduced for the full image reading, namely the removal of
temporary buffers where possible, writing directly to a NumericVector
instance where possible, which is then returned to the user without an
extra copy.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>